p_e/12: Don't actually care what the factors are
I don’t actually care what the factors are. I just care how many of
them there are. Instead of making an array, filling it, and then
getting its length, I’m making a count and incrementing it for each
factor found.

Unfortunately, this is only a shaved a percentage or two off of the
actual completion time. Probably more memory efficient though!