Project Euler Problem 23
Originally tried iterating through integers and testing sums for each
one of them - this crashed before it reached 30 and I started over.

I spent a lot of time on this one doing little optimizations that
ultimately cut the computation time significantly (including cutting
the time to generate abundant_sums by over 98%. However, this still
takes 95 seconds to run on my MBP, so I’ll be coming back to it to make
further adjustments.